backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langchain-pinecone
openai
pinecone
pypdf
python-dotenv